Translation of query.countdown.js
I have added a variable $lang, and using the wordpress function
get_locale I register and enqueue an additional script (the spanish
script of jquery.countdown.js).

The commit shows other changes I haven’t made, some code lines removed.
I suppose that it is because I have used the code of the plugin loaded
in the wordpress plugins page which must be different of the code you
have uploaded in github.