Add filters to adjust expiration time dynamically
Most likely to be used is 'wc_csr_expire_time_text' which adjusts the text string (ie: '5 Minutes').  This happens early enough that a specific item could be set to 'never' in case you don't want it
to expire or you could check if the user is logged in and force a specific expiration for those users.

The other filter is 'wc_csr_expire_time' which is used to adjust the number of seconds since the epoch for an item that will expire.  This is probably most useful if you want to do some kind of math
on the number rather than using plain english.  This happens after the 'never' check so if an item is set not to expire this filter will not be called

Fixes #51