Make sure cart items exist before trying to remove them
Related to #54

What is happening here is that if the WooCommerce Composite Product plugin is used and the composite item is marked as managed, then the composite item is removed first which also automatically
removes the child items.  When this plugin gets around to expiring the child items they are already removed from the cart and the error is generated.